# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  devmod.Access:
    properties:
      allow:
        $ref: '#/definitions/devmod.Allow'
      source:
        $ref: '#/definitions/devmod.Source'
    type: object
  devmod.Allow:
    properties:
      create:
        type: boolean
      delete:
        type: boolean
      read:
        type: boolean
      update:
        type: boolean
    type: object
  devmod.Device:
    properties:
      access:
        $ref: '#/definitions/devmod.Access'
      ataWsFlvUrl:
        type: string
      brand:
        type: string
      createAt:
        type: string
      district:
        type: string
      groupIds:
        items:
          type: string
        type: array
      id:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      password:
        type: string
      public:
        type: boolean
      roi:
        items:
          items:
            additionalProperties:
              type: string
            type: object
          type: array
        type: array
      state:
        type: string
      status:
        type: boolean
      updateAt:
        type: string
      url:
        type: string
      user:
        type: string
    type: object
...
  devmod.Source:
    properties:
      matchedGroupIds:
        items:
          type: string
        type: array
      mode:
        type: string
      subject:
        $ref: '#/definitions/devmod.Subject'
    type: object
  devmod.Subject:
    properties:
      id:
        type: string
      type:
        type: string
    type: object
...
  gmod.ApiErrorResponse:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details: {}
      message:
        example: Invalid request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  gmod.Pagination:
    properties:
      page:
        type: integer
      perPages:
        type: integer
      sortField:
        type: string
      sortOrder:
        type: string
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
  gmod.PaginationResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/devmod.Device'
        type: array
      pagination:
        $ref: '#/definitions/gmod.Pagination'
      status:
        example: true
        type: boolean
    type: object
...
  gmod.SuccessDataResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: REST API for Gateway system
  title: Gateway API
  version: "1.0"
paths:
...
  /ingest/unknownPayloadReviews:
    get:
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Per page
        in: query
        name: perPages
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.PaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: List unknown payload reviews
      tags:
      - ingest-unknown-payload-reviews
  /ingest/unknownPayloadReviews/{id}:
    get:
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDataResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unknown payload review
      tags:
      - ingest-unknown-payload-reviews
  /ingest/unknownPayloadReviews/{id}/reject:
    post:
      parameters:
      - description: Active Org ID
        in: header
        name: X-Active-Org
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gmod.SuccessDataResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gmod.ApiErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject unknown payload review and create a rejected payload pattern
      tags:
      - ingest-unknown-payload-reviews
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"